update and reduce databundle size
