+++ config/config.default.yaml
@@ -69,13 +69,10 @@ enable:
   retrieve: auto
   prepare_links_p_nom: false
   retrieve_databundle: true
-  retrieve_sector_databundle: true
   retrieve_cost_data: true
   build_cutout: false
   retrieve_irena: false
   retrieve_cutout: true
-  build_natura_raster: false
-  retrieve_natura_raster: true
   custom_busmap: false
   drop_leap_day: true
