+++ data/ch_cantons.csv
@@ -1,0 +1,27 @@
+Canton,HASC,NUTS
+Aargau,CH.AG,CH033
+Appenzell Inner Rhodes,CH.AI,CH054
+Appenzell Outer Rhodes,CH.AR,CH053
+Basel-Landschaft,CH.BL,CH032
+Basel-Stadt,CH.BS,CH031
+Bern,CH.BE,CH021
+Fribourg,CH.FR,CH022
+Geneva,CH.GE,CH013
+Glarus,CH.GL,CH051
+Graubünden,CH.GR,CH056
+Jura,CH.JU,CH025
+Lucerne,CH.LU,CH061
+Neuchâtel,CH.NE,CH024
+Nidwalden,CH.NW,CH065
+Obwalden,CH.OW,CH064
+Sankt Gallen,CH.SG,CH055
+Schaffhausen,CH.SH,CH052
+Schwyz,CH.SZ,CH063
+Solothurn,CH.SO,CH023
+Thurgau,CH.TG,CH057
+Ticino,CH.TI,CH07
+Uri,CH.UR,CH062
+Valais,CH.VS,CH012
+Vaud,CH.VD,CH011
+Zug,CH.ZG,CH066
+Zurich,CH.ZH,CH04
