+++ data/hydro_capacities.csv
@@ -1,0 +1,31 @@
+Country,p_nom_discharge[GW],p_nom_store[GW],E_store[TWh],InflowHourlyAvg[GWh]
+AT,13.08,3.8,3.2,4.02
+BE,1.42,1.31,0,0.04
+BA,2.05,0.62,2.5,0.71
+BG,3.13,0.86,4,0.53
+HR,2,0.61,2.8,0.57
+CZ,2.21,0.68,1.5,0.24
+DK,0.01,0,0,0
+EE,0.01,0,0,0
+FI,3.2,0,5.5,1.59
+FR,25.37,6.99,9.8,7.82
+DE,11.26,6.8,0.3,1.93
+GB,4.43,2.74,0,0.46
+GR,3.24,0.7,2.3,0.26
+HU,0.06,0,0.1,0.02
+IE,0.53,0.29,0,0.08
+IT,21.88,7.55,7.9,5.19
+LV,1.58,0,1.8,0.3
+LT,0.88,0.76,0.2,0.05
+LU,1.13,1.29,0,0
+NL,0.04,0,0,0.01
+NO,30.51,1.35,84.4,14
+PL,2.35,1.4,1.6,0.23
+PT,5.72,1.03,2.6,1.37
+RO,6.55,0.09,12.1,1.95
+RS,2.14,0.61,0,1.18
+SK,2.52,0.92,2.2,0.49
+SI,1.25,0.18,2.2,0.36
+ES,18.55,2.75,18.4,2.61
+SE,16.41,0.1,33.8,7.8
+CH,13.3,4.03,8.4,4.29
