+++ doc/preparation.rst
@@ -27,7 +27,6 @@ With these and the externally extracted ENTSO-E online map topology
 Then the process continues by calculating conventional power plant capacities, potentials, and per-unit availability time series for variable renewable energy carriers and hydro power plants with the following rules:
 
 - :mod:`build_powerplants` for today's thermal power plant capacities using `powerplantmatching <https://github.com/FRESNA/powerplantmatching>`__ allocating these to the closest substation for each powerplant,
-- :mod:`build_natura_raster` for rasterising NATURA2000 natural protection areas,
 - :mod:`build_ship_raster` for building shipping traffic density,
 - :mod:`build_renewable_profiles` for the hourly capacity factors and installation potentials constrained by land-use in each substation's Voronoi cell for PV, onshore and offshore wind, and
 - :mod:`build_hydro_profile` for the hourly per-unit hydro power availability time series.
@@ -54,14 +53,6 @@ Rule ``prepare_links_p_nom``
 ===============================
 
 .. automodule:: prepare_links_p_nom
-
-.. _natura:
-
-Rule ``build_natura_raster``
-===============================
-
-.. automodule:: build_natura_raster
-
 
 .. _base:
