+++ doc/retrieve.rst
@@ -53,32 +53,6 @@ The :ref:`tutorial` uses a smaller cutout than required for the full model (30 M
     For details see :mod:`build_cutout` and read the `atlite documentation <https://atlite.readthedocs.io>`__.
 
 
-Rule ``retrieve_natura_raster``
-================================
-
-.. image:: https://zenodo.org/badge/DOI/10.5281/zenodo.4706686.svg
-   :target: https://doi.org/10.5281/zenodo.4706686
-
-This rule, as a substitute for :mod:`build_natura_raster`, downloads an already rasterized version (`natura.tiff <https://zenodo.org/record/4706686/files/natura.tiff>`__) of `Natura 2000 <https://en.wikipedia.org/wiki/Natura_2000>`__ natural protection areas to reduce computation times. The file is placed into the ``resources`` sub-directory.
-
-**Relevant Settings**
-
-.. code:: yaml
-
-    enable:
-        build_natura_raster:
-
-.. seealso::
-    Documentation of the configuration file ``config/config.yaml`` at
-    :ref:`toplevel_cf`
-
-**Outputs**
-
-- ``resources/natura.tiff``: Rasterized version of `Natura 2000 <https://en.wikipedia.org/wiki/Natura_2000>`__ natural protection areas to reduce computation times.
-
-.. seealso::
-    For details see :mod:`build_natura_raster`.
-
 
 Rule ``retrieve_electricity_demand``
 ====================================
@@ -135,14 +109,3 @@ None.
 **Outputs**
 
 - ``data/shipdensity_global.zip``
-
-
-Rule ``retrieve_sector_databundle``
-====================================
-
-.. image:: https://zenodo.org/badge/DOI/10.5281/zenodo.5546516.svg
-   :target: https://doi.org/10.5281/zenodo.5546516
-
-In addition to the databundle required for electricity-only studies,
-another databundle is required for modelling sector-coupled systems.
-The size of this data bundle is around 640 MB.
