+++ rules/build_electricity.smk
@@ -109,7 +109,7 @@ rule build_shapes:
         nuts3=ancient("data/bundle/NUTS_2013_60M_SH/data/NUTS_RG_60M_2013.shp"),
         nuts3pop=ancient("data/bundle/nama_10r_3popgdp.tsv.gz"),
         nuts3gdp=ancient("data/bundle/nama_10r_3gdp.tsv.gz"),
-        ch_cantons=ancient("data/bundle/ch_cantons.csv"),
+        ch_cantons=ancient("data/ch_cantons.csv"),
         ch_popgdp=ancient("data/bundle/je-e-21.03.02.xls"),
     output:
         country_shapes=resources("country_shapes.geojson"),
@@ -172,27 +172,6 @@ if config["enable"].get("build_cutout", False):
             "../scripts/build_cutout.py"
 
 
-if config["enable"].get("build_natura_raster", False):
-
-    rule build_natura_raster:
-        input:
-            natura=ancient("data/bundle/natura/Natura2000_end2015.shp"),
-            cutout=lambda w: "cutouts/"
-            + CDIR
-            + config_provider("atlite", "default_cutout")(w)
-            + ".nc",
-        output:
-            resources("natura.tiff"),
-        resources:
-            mem_mb=5000,
-        log:
-            logs("build_natura_raster.log"),
-        conda:
-            "../envs/environment.yaml"
-        script:
-            "../scripts/build_natura_raster.py"
-
-
 rule build_ship_raster:
     input:
         ship_density="data/shipdensity_global.zip",
@@ -220,7 +199,7 @@ rule determine_availability_matrix_MD_UA:
         wdpa="data/WDPA.gpkg",
         wdpa_marine="data/WDPA_WDOECM_marine.gpkg",
         gebco=lambda w: (
-            "data/bundle/GEBCO_2014_2D.nc"
+            "data/bundle/gebco/GEBCO_2014_2D.nc"
             if config_provider("renewable", w.technology)(w).get("max_depth")
             else []
         ),
@@ -276,7 +255,7 @@ rule build_renewable_profiles:
         base_network=resources("networks/base.nc"),
         corine=ancient("data/bundle/corine/g250_clc06_V18_5.tif"),
         natura=lambda w: (
-            resources("natura.tiff")
+            "data/bundle/natura/natura.tiff"
             if config_provider("renewable", w.technology, "natura")(w)
             else []
         ),
@@ -287,7 +266,7 @@ rule build_renewable_profiles:
         ),
         gebco=ancient(
             lambda w: (
-                "data/bundle/GEBCO_2014_2D.nc"
+                "data/bundle/gebco/GEBCO_2014_2D.nc"
                 if config_provider("renewable", w.technology)(w).get("max_depth")
                 else []
             )
@@ -437,7 +416,7 @@ rule add_electricity:
         ),
         regions=resources("regions_onshore.geojson"),
         powerplants=resources("powerplants.csv"),
-        hydro_capacities=ancient("data/bundle/hydro_capacities.csv"),
+        hydro_capacities=ancient("data/hydro_capacities.csv"),
         geth_hydro_capacities="data/geth2015_hydro_capacities.csv",
         unit_commitment="data/unit_commitment.csv",
         fuel_price=lambda w: (
