+++ rules/build_sector.smk
@@ -287,10 +287,10 @@ rule build_energy_totals:
         energy=config_provider("energy"),
     input:
         nuts3_shapes=resources("nuts3_shapes.geojson"),
-        co2="data/bundle-sector/eea/UNFCCC_v23.csv",
+        co2="data/bundle/eea/UNFCCC_v23.csv",
         swiss="data/switzerland-new_format-all_years.csv",
         swiss_transport="data/gr-e-11.03.02.01.01-cc.csv",
-        idees="data/bundle-sector/jrc-idees-2015",
+        idees="data/bundle/jrc-idees-2015",
         district_heat_share="data/district_heat_share.csv",
         eurostat="data/eurostat/eurostat-energy_balances-april_2023_edition",
     output:
@@ -338,10 +338,10 @@ rule build_biomass_potentials:
             "https://zenodo.org/records/10356004/files/ENSPRESO_BIOMASS.xlsx",
             keep_local=True,
         ),
-        nuts2="data/bundle-sector/nuts/NUTS_RG_10M_2013_4326_LEVL_2.geojson",  # https://gisco-services.ec.europa.eu/distribution/v2/nuts/download/#nuts21
+        nuts2="data/bundle/nuts/NUTS_RG_10M_2013_4326_LEVL_2.geojson",  # https://gisco-services.ec.europa.eu/distribution/v2/nuts/download/#nuts21
         regions_onshore=resources("regions_onshore_elec_s{simpl}_{clusters}.geojson"),
         nuts3_population=ancient("data/bundle/nama_10r_3popgdp.tsv.gz"),
-        swiss_cantons=ancient("data/bundle/ch_cantons.csv"),
+        swiss_cantons=ancient("data/ch_cantons.csv"),
         swiss_population=ancient("data/bundle/je-e-21.03.02.xls"),
         country_shapes=resources("country_shapes.geojson"),
     output:
@@ -416,7 +416,7 @@ rule build_sequestration_potentials:
 
 rule build_salt_cavern_potentials:
     input:
-        salt_caverns="data/bundle-sector/h2_salt_caverns_GWh_per_sqkm.geojson",
+        salt_caverns="data/bundle/h2_salt_caverns_GWh_per_sqkm.geojson",
         regions_onshore=resources("regions_onshore_elec_s{simpl}_{clusters}.geojson"),
         regions_offshore=resources("regions_offshore_elec_s{simpl}_{clusters}.geojson"),
     output:
@@ -436,7 +436,7 @@ rule build_salt_cavern_potentials:
 
 rule build_ammonia_production:
     input:
-        usgs="data/bundle-sector/myb1-2017-nitro.xls",
+        usgs="data/bundle/myb1-2017-nitro.xls",
     output:
         ammonia_production=resources("ammonia_production.csv"),
     threads: 1
@@ -458,7 +458,7 @@ rule build_industry_sector_ratios:
         ammonia=config_provider("sector", "ammonia", default=False),
     input:
         ammonia_production=resources("ammonia_production.csv"),
-        idees="data/bundle-sector/jrc-idees-2015",
+        idees="data/bundle/jrc-idees-2015",
     output:
         industry_sector_ratios=resources("industry_sector_ratios.csv"),
     threads: 1
@@ -508,7 +508,7 @@ rule build_industrial_production_per_country:
         countries=config_provider("countries"),
     input:
         ammonia_production=resources("ammonia_production.csv"),
-        jrc="data/bundle-sector/jrc-idees-2015",
+        jrc="data/bundle/jrc-idees-2015",
         eurostat="data/eurostat/eurostat-energy_balances-april_2023_edition",
     output:
         industrial_production_per_country=resources(
@@ -564,7 +564,10 @@ rule build_industrial_distribution_key:
     input:
         regions_onshore=resources("regions_onshore_elec_s{simpl}_{clusters}.geojson"),
         clustered_pop_layout=resources("pop_layout_elec_s{simpl}_{clusters}.csv"),
-        hotmaps_industrial_database="data/bundle-sector/Industrial_Database.csv",
+        hotmaps_industrial_database=storage(
+            "https://gitlab.com/hotmaps/industrial_sites/industrial_sites_Industrial_Database/-/blob/master/data/Industrial_Database.csv",
+            keep_local=True,
+        ),
     output:
         industrial_distribution_key=resources(
             "industrial_distribution_key_elec_s{simpl}_{clusters}.csv"
@@ -652,7 +655,7 @@ rule build_industrial_energy_demand_per_country_today:
         countries=config_provider("countries"),
         industry=config_provider("industry"),
     input:
-        jrc="data/bundle-sector/jrc-idees-2015",
+        jrc="data/bundle/jrc-idees-2015",
         industrial_production_per_country=resources(
             "industrial_production_per_country.csv"
         ),
@@ -704,7 +707,7 @@ rule build_retro_cost:
         countries=config_provider("countries"),
     input:
         building_stock="data/retro/data_building_stock.csv",
-        data_tabula="data/bundle-sector/retro/tabula-calculator-calcsetbuilding.csv",
+        data_tabula="data/bundle/retro/tabula-calculator-calcsetbuilding.csv",
         air_temperature=resources("temp_air_total_elec_s{simpl}_{clusters}.nc"),
         u_values_PL="data/retro/u_values_poland.csv",
         tax_w="data/retro/electricity_taxes_eu.csv",
@@ -780,8 +783,8 @@ rule build_transport_demand:
             "pop_weighted_energy_totals_s{simpl}_{clusters}.csv"
         ),
         transport_data=resources("transport_data.csv"),
-        traffic_data_KFZ="data/bundle-sector/emobility/KFZ__count",
-        traffic_data_Pkw="data/bundle-sector/emobility/Pkw__count",
+        traffic_data_KFZ="data/bundle/emobility/KFZ__count",
+        traffic_data_Pkw="data/bundle/emobility/Pkw__count",
         temp_air_total=resources("temp_air_total_elec_s{simpl}_{clusters}.nc"),
     output:
         transport_demand=resources("transport_demand_s{simpl}_{clusters}.csv"),
@@ -925,7 +928,7 @@ rule prepare_sector_network:
         avail_profile=resources("avail_profile_s{simpl}_{clusters}.csv"),
         dsm_profile=resources("dsm_profile_s{simpl}_{clusters}.csv"),
         co2_totals_name=resources("co2_totals.csv"),
-        co2="data/bundle-sector/eea/UNFCCC_v23.csv",
+        co2="data/bundle/eea/UNFCCC_v23.csv",
         biomass_potentials=lambda w: (
             resources(
                 "biomass_potentials_s{simpl}_{clusters}_"
