+++ rules/postprocess.smk
@@ -230,7 +230,7 @@ rule plot_summary:
         energy=RESULTS + "csvs/energy.csv",
         balances=RESULTS + "csvs/supply_energy.csv",
         eurostat="data/eurostat/eurostat-energy_balances-april_2023_edition",
-        co2="data/bundle-sector/eea/UNFCCC_v23.csv",
+        co2="data/bundle/eea/UNFCCC_v23.csv",
     output:
         costs=RESULTS + "graphs/costs.pdf",
         energy=RESULTS + "graphs/energy.pdf",
