+++ scripts/build_ship_raster.py
@@ -47,9 +47,34 @@ from pathlib import Path
 
 import rioxarray
 from _helpers import configure_logging, set_scenario_config
-from build_natura_raster import determine_cutout_xXyY
 
 logger = logging.getLogger(__name__)
+
+
+def determine_cutout_xXyY(cutout_name):
+    """
+    Determine the full extent of a cutout.
+
+    Since the coordinates of the cutout data are given as the
+    center of the grid cells, the extent of the cutout is
+    calculated by adding/subtracting half of the grid cell size.
+
+
+    Parameters
+    ----------
+    cutout_name : str
+        Path to the cutout.
+
+    Returns
+    -------
+    A list of extent coordinates in the order [x, X, y, Y].
+    """
+    cutout = atlite.Cutout(cutout_name)
+    assert cutout.crs.to_epsg() == 4326
+    x, X, y, Y = cutout.extent
+    dx, dy = cutout.dx, cutout.dy
+    return [x - dx / 2.0, X + dx / 2.0, y - dy / 2.0, Y + dy / 2.0]
+
 
 if __name__ == "__main__":
     if "snakemake" not in globals():
