+++ scripts/retrieve_databundle.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright 2019-2022 Fabian Hofmann (TUB, FIAS)
+# Copyright 2019-2024 Fabian Hofmann (TUB, FIAS), Fabian Neumann (TUB)
 # SPDX-FileCopyrightText: : 2017-2024 The PyPSA-Eur Authors
 #
 # SPDX-License-Identifier: MIT
@@ -7,24 +7,15 @@
 .. image:: https://zenodo.org/badge/DOI/10.5281/zenodo.3517935.svg
    :target: https://doi.org/10.5281/zenodo.3517935
 
-The data bundle (1.4 GB) contains common GIS datasets like NUTS3 shapes, EEZ shapes, CORINE Landcover, Natura 2000 and also electricity specific summary statistics like historic per country yearly totals of hydro generation, GDP and POP on NUTS3 levels and per-country load time-series.
+The data bundle contains common GIS datasets like NUTS3 shapes, EEZ shapes,
+CORINE Landcover, Natura 2000 and also electricity specific summary statistics
+like historic per country yearly totals of hydro generation, GDP and population
+data on NUTS3 levels and energy balances.
 
-This rule downloads the data bundle from `zenodo <https://doi.org/10.5281/zenodo.3517935>`_ and extracts it in the ``data`` sub-directory, such that all files of the bundle are stored in the ``data/bundle`` subdirectory.
-
-The :ref:`tutorial` uses a smaller `data bundle <https://zenodo.org/record/3517921/files/pypsa-eur-tutorial-data-bundle.tar.xz>`_ than required for the full model (188 MB)
-
-.. image:: https://zenodo.org/badge/DOI/10.5281/zenodo.3517921.svg
-    :target: https://doi.org/10.5281/zenodo.3517921
-
-**Relevant Settings**
-
-.. code:: yaml
-
-    tutorial:
-
-.. seealso::
-    Documentation of the configuration file ``config/config.yaml`` at
-    :ref:`toplevel_cf`
+This rule downloads the data bundle from `zenodo
+<https://doi.org/10.5281/zenodo.3517935>`_ and extracts it in the ``data``
+sub-directory, such that all files of the bundle are stored in the
+``data/bundle`` subdirectory.
 
 **Outputs**
 
@@ -57,10 +48,7 @@ if __name__ == "__main__":
     configure_logging(snakemake)
     set_scenario_config(snakemake)
 
-    if snakemake.config["tutorial"]:
-        url = "https://zenodo.org/record/3517921/files/pypsa-eur-tutorial-data-bundle.tar.xz"
-    else:
-        url = "https://zenodo.org/record/3517935/files/pypsa-eur-data-bundle.tar.xz"
+    url = "https://zenodo.org/records/10973944/files/bundle.tar.xz"
 
     tarball_fn = Path(f"{rootpath}/bundle.tar.xz")
     to_fn = Path(rootpath) / Path(snakemake.output[0]).parent.parent
@@ -74,6 +62,7 @@ if __name__ == "__main__":
     logger.info("Extracting databundle.")
     tarfile.open(tarball_fn).extractall(to_fn)
 
+    logger.info("Unlinking tarball.")
     tarball_fn.unlink()
 
     logger.info(f"Databundle available in '{to_fn}'.")
